requirements.txt
python-dotenv==0.21.0
google-api-python-client==2.162.0
APScheduler==3.11.0
google-genai==1.3.0
pylovepdf==1.3.2
yfinance==0.2.54